full_requirements.txt
torch==1.7.0
torch-model-archiver==0.3.0
torchserve==0.3.0
torchsummary==1.5.1
torchtext==0.8.0
Cython==0.29.21
scikit-image==0.16.2
scipy==1.4.1
pyyaml==5.1 
yacs==0.1.8
urllib3==1.25.4
farm-haystack==0.8.0
transformers==4.1.1
datasets==1.6.2
